Add alliance wrapper and event
This alliance wrapper hopefully negates any pain that would come up with using it,
since the name of the obfuscated class is equal to one of a package,
the compiler will refuse to compile any references to that class if it is not performed
via Reflection. Luckily Mixins allow to mixin in classes without a class reference, otherwise
this would have been impossible to do.