Expanded Actor specification API
It still isn't really usable and even the spacecrafts spec API isn't
polished as of yet, for example it is clear that more Events are
required to deal with the actions of the actors.

I also have not yet tested on how the Mixin framework reacts about our
hierarchy, but I doubt it would not work. If it does not work then that
would be a significant pain to resolve.
Another doubt is whether the Spacecraft specs default interfaces override
the default method provided by the StateActorMixins.