Create a way to create Screen instances
Now there is the need of obtaining ScreenComponents to solve and then we
might have finally a decent GUI implementation that is more complex than
the dialog system that was introduced in 1.1.

It's kinda strange to know that SLAPI 1.1 wasn't yesterday. Yet I
believe that I have done almost nothing since then. Obviously I am
wrong, but oh well.