Properly downgrade GDX
I have the sneaking suspicion that these libraries were never updated
after the inital release of galimulator. We might have to do this ourselves
as at some point they will refuse to work in newer java versions.

But perhaps Windows 11 might nudge the developer to update the libs, as
apparently galimulator does not work correctly for some on this OS