Move to our fork of brachyura (slbrachyura) and Starplane.
The second (or third?) generation of galimulator build tools is out.

It remains to be seen whether it will be the last generation or
whether there is a successor to it. It also remains to be seen
on how much of a good idea forking brachyura was. It is certainly
less reliable but at least have some easily extendable API and forking
it was certainly a better idea than staying with upstream.
Upstream will also probably not recieve our improvements as
they are pretty radical in many areas.

However automatic mapping of deobfuscated names will probably be the
core process of our build supply chain. Even if in the end manual
mapping may be the be-all and end-all. But given the current size
of the galimulator modding community (1 person as of right now) it
is probably just a waste of man hours for now. In the future it might
be more attractive as I might do my "Besondere Lernleistung" on
automatic updating of deobfuscation mappings however such a project
will take a lot of time. As such I might think that selling the
Starloader-project as the "Besondere Lernleistung" might be more viable,
especially because I have now worked over a year on it and at the time that
I will have to hand over the "Besondere Lernleistung" I would have
commited around 3 years and probably over 4 000 hours to the entirety
of the project. Numbers which are unthinkable for my peers. But that is
what you can do if you have no social life.