Removed large parts of the actor spec API and other refractors
This is basically the first of a series of very large refractors
that have the goal of improving the longevity of the project.
Many APIs were initially developed for no reason, and the actor API
is one of them. Not only was it lackluster in practice, extending
the vanilla game was also pretty much impossible with it.
The goal will be to add a JSON-centric actor API while still allowing to
spawn most (if not all) built-in actors.

Furthermore this commit cuts down on usage of internal galimulator API.
References of GalColor have been removed and replaced with GDX's Color.
To fully embrace the spirit of that change, the TextColor enum was removed
and references to AWT's Color class was removed too. They were also replaced
the GDX's color class.
Additionally a really thin Fleet API was added as previously galimulator's
fleet class was referenced. This reference has been replaced with a reference
to the newbord fleet API. That fleet API will be expanded in the coming
updates.
Another change is the removal of references to the religion object, which
have been replaced with a namespaced key. To accompany this change the
religion enum was turned into a registry. The religion object may
recieve a more specific interface in the future