Savegame metadata API v1
There is the need for more codecs and overall a more polished API,
but this is at least usable for small-sized objects. For any objects
over 32k bytes in size, this is not really an option though I will eventually
search for one.

This system is also able to load vanilla saves, though right now the user
is not able to save vanilla saves though API-wise this is allowed.
The goal will be to allow the player to choose between savegame systems,
though this might be a bit annoying so I am not too sure about that.

Anyways, I am a bit surprised that I did not see the need of such an API
beforehand because quite honestly you need this for EVERYTHING.
I should also improve SpStarmap a bit more, there is an awful lot of unmapped
stuff used in the vanilla savegame format class.

Another thing that might be interesting to know is whether this API
is capable of actually surviving for long in the wild - while I have tried
to write it in a way that it is versatile, I am not really pleased with it
even if it largely met my expectations for the API. There are a lot of design
flaws but I don't think that you can get rid of them anytime soon. I should
also get rid of Optional getters as they are kindof prone to errors