Canvas API
Also adds a few QoL features such as the unofficial (impl-specific) Usertest API,
"gracefull" crashing and a few methods here and there.

The canvas API is more or less the long-awaited successor to the screen API,
which was deemed to be beyond repair and should get deprecated for removal shortly.
In the meantime a translation layer may get added to facilitate the move.

The reason the screen API was such a disappointment was that the screen API was only
really written for one specific purpose, which at the time was to display a string.
Only later on reactivity was added, which was kindof too late at that point.

Furthermore the screen API was only really viable to display a single component,
displaying more than one component is probably impossible to do without leaving
the land of sanity.

This behaviour was caused by a significant lack of understanding of galimulator's
and libGDX's drawing logic as well as a lack of testing. To prevent this from repeating,
the current unoffical usertest API was populated with several tests concerning canvases.

Another cause of this debacle was the fundemental design flaw of not wishing to interact
with galimulator's internal API. Meanwhile the canvas API is mostly interconnected with
the internal widget API, which improves usability significantly.

However with galimulator 5.0 (or 4.11?) looming across the horizon, such interconnection
can be harmfull as 5.0 may bring significant changes to the drawing logic. But the
potential for damage is a lot less in the canvas API because in the end it is a lot more
independent from the widget API, where as the screen API is a lot more dependent on
certain implementation specifics that are flawed.

Furthermore the added interactive tests will help resolve future incompatibilities
a lot quicker.

Another flaw of the screen API was the layout, though that one wasn't really visible
because you wouldn't want to add more than two components to a screen anyways.
The canvas API uses a simple single row/column based layout which makes it easier to
understand what is going on, albeit being a little bit more expensive on the performance
side.

In the end I am positive that the canvas API is the way forward until we get something better