Bugfixes, Prettier save/load UI
Concretely this commit fixes a bug where autosaves would likely not work
as they would happen concurrently to the tick loop. It also updates
starplane, so loading saves should now be possible in the dev environment.
That being said this update does have it's cons as the line numbers in
stacktraces are no longer reflected by those in the actual files.
But that should be rather easy to fix