Mark the Text API as deprecated and for removal.
In retrospect, the Text API was a mistake. It had an incredibly niche
use within the Bulletin system. That being said it wasn't given much
attention after it was concieved over 1 and a half years ago. Now it
seems to have broken to an unknown extent, at which point it was
decided to deprecate it and remove it later on should a larger change
happen that would result in larger maintainance work. Should no such
change arise, the Text API would get culled in the 3.0.0 refractor.