Update to latest 5.0 beta release.
Happy new year from the SLAPI team.
This project is now in it's third year. 2021 was the beginning, 2022 saw significant
improvements to the buildchain and 2023 will offer something unknown. Perhaps it will
be something new, perhaps something old. I cannot tell yet what it will be,
but I believe it will be a mix of both.