Fix compilation
Turns out the switch from buildscript files that were untouched for a better part of a year
to a mostly new build system wasn't as clean as first assumed. But this should be every
issue fixed and should only block on gslSP now. I think I'll commit the changes to
the v2.0 refractor branch once galimulator 5.0 gets fully released (or it is abundantly
clear that the 5.0 beta is the last galimulator update). With that I think I'll commit
the v2.0 branch into main and make it the first release since April 2022 (over a year ago!)