Migrate to stianloader.org maven, use CI
Small mishap on the gradle module metadata remapping part - I completely forgot about it.
Oh well, what's done is done and I'll be able to remove the related hacks in SLAPI
once the erroneous versions are no longer used. Thankfully this is now a nonissue on the
side of our CI.