Lift deprecation of IntegerChooseOption
Also reworded the documentation of IntegerChooseoption
and StringChooseOption to mitigate the risk of these interfaces
being erroneously deprecated.

However the proper fix would be to provide an implementation
for both of these classes, but at the moment that is not a pressing
concern and deprecating these interfaces due to there not being a
ready-made implementation available is not a good reason
for deprecating these interfaces for removal.