Allow to disable star render region transformation logic
Mods are now expected to reimplement the transformations
applied by SLAPI though, unless they do not touch the magic method.

This is more of a temporary workaround though, in the long run I should
be looking at whether there are other ways of implementing these transformations
or handling these transformation failures more gracefully.

But as time will most likely tell, this workaround will not be
that temporary