Refractor the ActiveEmpire interface into the Empire interface
This will be a multi-stage refractor with this commit being the bulk of the changes for now.
Later on (in SLAPI 3X), the old Empire interface will be removed.

And with this one of the last blockers on a SLAPI 2.0.0 full release is released,
so perhaps that will happen in the not so near future - who knows?