Fix concurrency crash caused by having the profiler open
Note that this does not fix the crash that occurs when opening the profiler.
That too is a concurrency issue and it can only really be mitigated using a
mixin with an Inject at HEAD and using cancellable = true. However,
right now I do not believe that is necessary and this change should suffice
to prevent the game from crashing for most players.