Partially update README file
It was a bit dated. The example page is a bit dated too,
but I'm a bit exhausted right now with the fact that I had
my "abiturfeier" today and talked for a bit longer with certain
people than would be good for my own benefit.

Or perhaps I'm wrong - it's a bit difficult to say in the heat
of the moment. Thankfully my past me did the decision to
pretty much promise to write a farewell message to someone
who is dear to me, but I believe that ultimately such a message
will have a neutral outcome (i.e. no change). Perhaps that's
a good outcome?

Also, perhaps I should stop spordically using Github as
my own version of a diary. But considering that I'll likely
be the sole person reviewing the commits (galimulator is a
graveyard at this point) - what gives at the end of the day?