Fix crash during galaxy generation when no natives are registered
This is generally not the case, but at times people may wish to
play without natives. In vanilla galimulator this would yield a
`java.lang.IllegalArgumentException: n must be positive` from the
RandomXS128 class, but this commit fixes this behaviour in modded
galimulator if SLAPI is being used (so pretty much all cases).