Deprecate NullUtils#requireNotNull for removal
The entire class is expected to be deprecated for removal later on,
however this should mark the biggest change. The official replacement
is Objects#requireNonNull. The NullUtils class is mostly used by older
code, so this commit contains a lot of small style changes in or near
affected code ontop of removing references to the deprecated method.