Add Console prefab component
Slowly, very slowly, the prefabs come rolling in.
Ideally this will mean that writing GUIs becomes easier as time marches on,
but honestly GUIs will probably forever be a sore point. At least for as long
as noone else joins in on the fun, but I can't see that ever happening.

Quite sad, pretty much? Well, I donno.

Regardless, a Console should be one of the more versatile tools
out there. It could be used for chatting, a command terminal, and
that's probably all the uses I can forsee. I guess it's also fit as
a glorified text box?