Fix LogicalTickEvent not firing
I mean, it's fine because that event was last used in the Timelapser mod which
wasn't that much of a best-selling mod (who even remembers it now other than me
lol?), but still ... wow, to have a bug that is kinda critical for many mods
stay for roundabout 2 years (maybe even 3, I donno and I won't look it up)
is quite frankly insane.

Well, here's to that this will never happen again.