Fix Net#openURI implementations on linux
In LWJGL2, the implementation would cause log spam when sensible-browser is not installed,
while in LWJGL3 installations, if your AWT implementation is not working as it should be,
openURI would outright not work. Instead exceptions are now supressed and only logged if
all options are exhausted, and on LWJGL3 xdg-open will be directly invoked if AWT does not
support Desktop#browse for one reason or another.