Fix severe performance issues induced by actor dragging
Also introduced a few useless classes (namely the Atomics class),
but it could be useful in a few years so it'll stay. Tech debt for the win.

I also had to introduce a wrapper around the ActorDragManager constructor
because javac decided to no longer compile that class??? Whatever,
this is not the first time we've seen inner classes suddenly no longer
being handleable by javac so thankfully the majority of the infrastructure
for that is already in place.